+++ geolocator_apple/ios/Classes/Handlers/LocationAccuracyHandler.h
@@ -1,0 +1,1 @@
+../../../apple/Classes/Handlers/LocationAccuracyHandler.h