+++ geolocator_apple/ios/Classes/Handlers/PermissionHandler.m
@@ -1,0 +1,1 @@
+../../../apple/Classes/Handlers/PermissionHandler.m