+++ geolocator_apple/ios/Classes/Handlers/PositionStreamHandler.h
@@ -1,0 +1,1 @@
+../../../apple/Classes/Handlers/PositionStreamHandler.h