+++ geolocator_apple/macos/Classes/Constants/ErrorCodes.h
@@ -1,0 +1,1 @@
+../../../apple/Classes/Constants/ErrorCodes.h