+++ geolocator_apple/macos/Classes/Handlers/GeolocationHandler.m
@@ -1,0 +1,1 @@
+../../../apple/Classes/Handlers/GeolocationHandler.m