+++ geolocator_apple/macos/Classes/Utils/AuthorizationStatusMapper.h
@@ -1,0 +1,1 @@
+../../../apple/Classes/Utils/AuthorizationStatusMapper.h