+++ geolocator_platform_interface/test/src/enums/location_accuracy_test.dart
@@ -2,14 +2,14 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:geolocator_platform_interface/geolocator_platform_interface.dart';
 
 void main() {
-  test('LocationAccuracy should contain 6 options', () {
-    final values = LocationAccuracy.values;
+  test('LocationAccuracy should contain 7 options', () {
+    const values = LocationAccuracy.values;
 
-    expect(values.length, 6);
+    expect(values.length, 7);
   });
 
   test("LocationAccuracy enum should have items in correct index", () {
-    final values = LocationAccuracy.values;
+    const values = LocationAccuracy.values;
 
     expect(values[0], LocationAccuracy.lowest);
     expect(values[1], LocationAccuracy.low);
@@ -17,5 +17,6 @@ void main() {
     expect(values[3], LocationAccuracy.high);
     expect(values[4], LocationAccuracy.best);
     expect(values[5], LocationAccuracy.bestForNavigation);
+    expect(values[6], LocationAccuracy.reduced);
   });
 }