+++ geolocator_platform_interface/test/src/errors/already_subscribed_exception_test.dart
@@ -4,10 +4,10 @@ import 'package:geolocator_platform_interface/geolocator_platform_interface.dart
 void main() {
   test('toString: Should return a description of the exception', () {
     // Arrange
-    final expected = 'The App is already listening to a stream of position '
+    const expected = 'The App is already listening to a stream of position '
         'updates. It is not possible to listen to more then one stream at the '
         'same time.';
-    final exception = AlreadySubscribedException();
+    const exception = AlreadySubscribedException();
 
     // Act
     final actual = exception.toString();