+++ geolocator_platform_interface/test/src/implementations/method_channel_geolocator_test.dart
@@ -135,10 +135,90 @@ void main() {
       });
     });
 
+    group(
+        'requestTemporaryFullAccuracy: When requesting temporary full'
+        'accuracy.', () {
+      test(
+          'Should receive reduced accuracy if Location Accuracy is pinned to'
+          ' reduced', () async {
+        // Arrange
+        final methodChannel = MethodChannelMock(
+            channelName: 'flutter.baseflow.com/geolocator',
+            method: 'requestTemporaryFullAccuracy',
+            result: 0);
+
+        final expectedArguments = <String, dynamic>{
+          'purposeKey': 'purposeKeyValue',
+        };
+
+        // Act
+        final accuracy =
+            await MethodChannelGeolocator().requestTemporaryFullAccuracy(
+          purposeKey: 'purposeKeyValue',
+        );
+
+        // Assert
+        expect(accuracy, LocationAccuracyStatus.reduced);
+
+        expect(methodChannel.log, <Matcher>[
+          isMethodCall(
+            'requestTemporaryFullAccuracy',
+            arguments: expectedArguments,
+          ),
+        ]);
+      });
+
+      test(
+          'Should receive reduced accuracy if Location Accuracy is already set'
+          ' to precise location accuracy', () async {
+        // Arrange
+        MethodChannelMock(
+            channelName: 'flutter.baseflow.com/geolocator',
+            method: 'requestTemporaryFullAccuracy',
+            result: 1);
+
+        // Act
+        final accuracy = await MethodChannelGeolocator()
+            .requestTemporaryFullAccuracy(purposeKey: 'purposeKey');
+
+        // Assert
+        expect(accuracy, LocationAccuracyStatus.precise);
+      });
+
+      test('Should receive an exception when permission definitions not found',
+          () async {
+        // Arrange
+        MethodChannelMock(
+          channelName: 'flutter.baseflow.com/geolocator',
+          method: 'requestTemporaryFullAccuracy',
+          result: PlatformException(
+            code: 'PERMISSION_DEFINITIONS_NOT_FOUND',
+            message: 'Permission definitions are not found.',
+            details: null,
+          ),
+        );
+
+        // Act
+        final future = MethodChannelGeolocator()
+            .requestTemporaryFullAccuracy(purposeKey: 'purposeKey');
+
+        // Assert
+        expect(
+          future,
+          throwsA(
+            isA<PermissionDefinitionsNotFoundException>().having(
+              (e) => e.message,
+              'description',
+              'Permission definitions are not found.',
+            ),
+          ),
+        );
+      });
+    });
+
     group('getLocationAccuracy: When requesting the Location Accuracy Status',
         () {
-      test(
-          'Should receive reduced accuracy if Location Accuracy is reduced',
+      test('Should receive reduced accuracy if Location Accuracy is reduced',
           () async {
         // Arrange
         MethodChannelMock(
@@ -155,24 +235,22 @@ void main() {
         expect(locationAccuracy, LocationAccuracyStatus.reduced);
       });
 
-      test(
-          'Should receive reduced accuracy if Location Accuracy is reduced',
-              () async {
-            // Arrange
-            MethodChannelMock(
-              channelName: 'flutter.baseflow.com/geolocator',
-              method: 'getLocationAccuracy',
-              result: 1,
-            );
+      test('Should receive reduced accuracy if Location Accuracy is reduced',
+          () async {
+        // Arrange
+        MethodChannelMock(
+          channelName: 'flutter.baseflow.com/geolocator',
+          method: 'getLocationAccuracy',
+          result: 1,
+        );
 
-            // Act
-            final locationAccuracy =
+        // Act
+        final locationAccuracy =
             await MethodChannelGeolocator().getLocationAccuracy();
 
-            // Assert
-            expect(locationAccuracy, LocationAccuracyStatus.precise);
-          });
-
+        // Assert
+        expect(locationAccuracy, LocationAccuracyStatus.precise);
+      });
     });
 
     group('requestPermission: When requesting for permission', () {
@@ -443,7 +521,7 @@ void main() {
             channelName: 'flutter.baseflow.com/geolocator',
             method: 'getCurrentPosition',
             result: mockPosition.toJson());
-        final expectedArguments = LocationOptions(
+        const expectedArguments = LocationOptions(
           accuracy: LocationAccuracy.low,
         );
 
@@ -469,11 +547,11 @@ void main() {
           method: 'getCurrentPosition',
           result: mockPosition.toJson(),
         );
-        final expectedFirstArguments = LocationOptions(
+        const expectedFirstArguments = LocationOptions(
           accuracy: LocationAccuracy.low,
           forceAndroidLocationManager: false,
         );
-        final expectedSecondArguments = LocationOptions(
+        const expectedSecondArguments = LocationOptions(
           accuracy: LocationAccuracy.high,
           forceAndroidLocationManager: true,
         );
@@ -563,7 +641,7 @@ void main() {
         // Arrange
         MethodChannelMock(
           channelName: 'flutter.baseflow.com/geolocator',
-          delay: Duration(milliseconds: 10),
+          delay: const Duration(milliseconds: 10),
           method: 'getCurrentPosition',
           result: mockPosition.toJson(),
         );
@@ -572,7 +650,7 @@ void main() {
           await MethodChannelGeolocator().getCurrentPosition(
             desiredAccuracy: LocationAccuracy.low,
             forceAndroidLocationManager: true,
-            timeLimit: Duration(milliseconds: 5),
+            timeLimit: const Duration(milliseconds: 5),
           );
 
           fail('Expected a TimeoutException and should not reach here.');
@@ -596,25 +674,39 @@ void main() {
           );
         });
 
-        test('Should return a new stream when original stream is closed', () {
+        test('Should return a new stream when all subscriptions are cancelled',
+            () {
           final methodChannelGeolocator = MethodChannelGeolocator();
-          final firstStream =
-              methodChannelGeolocator.getPositionStream(distanceFilter: 10);
 
-          // Start listening so u can cancel the stream subscription
+          // Get two position streams
+          final firstStream = methodChannelGeolocator.getPositionStream();
+          final secondStream = methodChannelGeolocator.getPositionStream();
+
+          // Streams are the same object
+          expect(firstStream == secondStream, true);
+
+          // Add multiple subscriptions
           StreamSubscription<Position>? firstSubscription =
               firstStream.listen((event) {});
+          StreamSubscription<Position>? secondSubscription =
+              secondStream.listen((event) {});
 
-          // Cancel subscription
+          // Cancel first subscription
           firstSubscription.cancel();
           firstSubscription = null;
 
-          final secondStream =
-              methodChannelGeolocator.getPositionStream(distanceFilter: 100);
+          // Stream is still the same as the first one
+          final cachedStream = methodChannelGeolocator.getPositionStream();
+          expect(firstStream == cachedStream, true);
 
-          // Pro stream different options are used, thus the stream shouldn't
-          // be the same
-          expect(firstStream == secondStream, true);
+          // Cancel second subscription
+          secondSubscription.cancel();
+          secondSubscription = null;
+
+          // After all listeners have been removed, the next stream
+          // retrieved is a new one.
+          final thirdStream = methodChannelGeolocator.getPositionStream();
+          expect(firstStream != thirdStream, true);
         });
       });
 
@@ -644,7 +736,7 @@ void main() {
           'Should correctly handle done event', () async {
         // Arrange
         final completer = Completer();
-        completer.future.timeout(Duration(milliseconds: 50),
+        completer.future.timeout(const Duration(milliseconds: 50),
             onTimeout: () =>
                 fail('getPositionStream should trigger done and not timeout.'));
         final streamController =
@@ -876,7 +968,7 @@ void main() {
           channelName: 'flutter.baseflow.com/geolocator_updates',
           stream: streamController.stream,
         );
-        final expectedArguments = LocationOptions(
+        const expectedArguments = LocationOptions(
           accuracy: LocationAccuracy.low,
           distanceFilter: 0,
         );
@@ -884,12 +976,12 @@ void main() {
         // Act
         final positionStream = MethodChannelGeolocator().getPositionStream(
             desiredAccuracy: expectedArguments.accuracy,
-            timeLimit: Duration(milliseconds: 5));
+            timeLimit: const Duration(milliseconds: 5));
         final streamQueue = StreamQueue(positionStream);
 
         streamController.add(mockPosition.toJson());
 
-        await Future.delayed(Duration(milliseconds: 5));
+        await Future.delayed(const Duration(milliseconds: 5));
 
         // Assert
         expect(await streamQueue.next, mockPosition);