+++ geolocator_web/example/lib/main.dart
@@ -1,0 +1,252 @@
+import 'dart:async';
+
+import 'package:baseflow_plugin_template/baseflow_plugin_template.dart';
+import 'package:flutter/material.dart';
+import 'package:geolocator_web/geolocator_web.dart';
+
+/// Defines the main theme color.
+final MaterialColor themeMaterialColor =
+    BaseflowPluginExample.createMaterialColor(
+        const Color.fromRGBO(48, 49, 60, 1));
+
+void main() {
+  runApp(BaseflowPluginExample(
+    pluginName: 'Geolocator',
+    githubURL: 'https://github.com/Baseflow/flutter-geolocator',
+    pubDevURL: 'https://pub.dev/packages/geolocator',
+    pages: [GeolocatorWidget.createPage()],
+  ));
+}
+
+/// Example [Widget] showing the functionalities of the geolocator plugin
+class GeolocatorWidget extends StatefulWidget {
+  /// Create a GeolocatorWidget.
+  const GeolocatorWidget({Key? key}) : super(key: key);
+
+  /// Utility method to create a page with the Baseflow templating.
+  static ExamplePage createPage() {
+    return ExamplePage(
+        Icons.location_on, (context) => const GeolocatorWidget());
+  }
+
+  @override
+  _GeolocatorWidgetState createState() => _GeolocatorWidgetState();
+}
+
+class _GeolocatorWidgetState extends State<GeolocatorWidget> {
+  static const String _locationServicesDisabledMessage =
+      'Location services are disabled.';
+  static const String _permissionDeniedMessage = 'Permission denied.';
+  static const String _permissionDeniedForeverMessage =
+      'Permission denied forever.';
+  static const String _permissionGrantedMessage = 'Permission granted.';
+
+  final GeolocatorPlatform _geolocatorPlatform = GeolocatorPlatform.instance;
+  final List<_PositionItem> _positionItems = <_PositionItem>[];
+  StreamSubscription<Position>? _positionStreamSubscription;
+
+  @override
+  Widget build(BuildContext context) {
+    const buttonSpacer = SizedBox(
+      height: 10,
+    );
+
+    return Scaffold(
+      backgroundColor: Theme.of(context).backgroundColor,
+      body: ListView.builder(
+        itemCount: _positionItems.length,
+        itemBuilder: (context, index) {
+          final positionItem = _positionItems[index];
+
+          if (positionItem.type == _PositionItemType.log) {
+            return ListTile(
+              title: Text(positionItem.displayValue,
+                  textAlign: TextAlign.center,
+                  style: const TextStyle(
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                  )),
+            );
+          } else {
+            return Card(
+              child: ListTile(
+                tileColor: themeMaterialColor,
+                title: Text(
+                  positionItem.displayValue,
+                  style: const TextStyle(color: Colors.white),
+                ),
+              ),
+            );
+          }
+        },
+      ),
+      floatingActionButton: Column(
+        crossAxisAlignment: CrossAxisAlignment.end,
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: [
+          FloatingActionButton(
+            child: (_positionStreamSubscription == null ||
+                    _positionStreamSubscription!.isPaused)
+                ? const Icon(Icons.play_arrow)
+                : const Icon(Icons.pause),
+            onPressed: _toggleListening,
+            tooltip: (_positionStreamSubscription == null)
+                ? 'Start position updates'
+                : _positionStreamSubscription!.isPaused
+                    ? 'Resume'
+                    : 'Pause',
+            backgroundColor: _determineButtonColor(),
+          ),
+          buttonSpacer,
+          FloatingActionButton(
+            child: const Icon(Icons.my_location),
+            onPressed: _getCurrentPosition,
+          ),
+          buttonSpacer,
+          FloatingActionButton(
+            child: const Icon(Icons.close),
+            onPressed: () => setState(_positionItems.clear),
+            tooltip: 'clear',
+          ),
+        ],
+      ),
+    );
+  }
+
+  Future<void> _getCurrentPosition() async {
+    final bool hasPermission = await _handlePermission();
+
+    if (!hasPermission) {
+      return;
+    }
+
+    final Position position = await _geolocatorPlatform.getCurrentPosition();
+    _updatePositionList(
+      _PositionItemType.position,
+      position.toString(),
+    );
+  }
+
+  Future<bool> _handlePermission() async {
+    bool serviceEnabled;
+    LocationPermission permission;
+
+    // Test if location services are enabled.
+    serviceEnabled = await _geolocatorPlatform.isLocationServiceEnabled();
+    if (!serviceEnabled) {
+      // Location services are not enabled don't continue
+      // accessing the position and request users of the
+      // App to enable the location services.
+      _updatePositionList(
+        _PositionItemType.log,
+        _locationServicesDisabledMessage,
+      );
+
+      return false;
+    }
+
+    permission = await _geolocatorPlatform.checkPermission();
+    if (permission == LocationPermission.denied) {
+      permission = await _geolocatorPlatform.requestPermission();
+      if (permission == LocationPermission.denied) {
+        // Permissions are denied, next time you could try
+        // requesting permissions again (this is also where
+        // Android's shouldShowRequestPermissionRationale
+        // returned true. According to Android guidelines
+        // your App should show an explanatory UI now.
+        _updatePositionList(
+          _PositionItemType.log,
+          _permissionDeniedMessage,
+        );
+
+        return false;
+      }
+    }
+
+    if (permission == LocationPermission.deniedForever) {
+      // Permissions are denied forever, handle appropriately.
+      _updatePositionList(
+        _PositionItemType.log,
+        _permissionDeniedForeverMessage,
+      );
+
+      return false;
+    }
+
+    // When we reach here, permissions are granted and we can
+    // continue accessing the position of the device.
+    _updatePositionList(
+      _PositionItemType.log,
+      _permissionGrantedMessage,
+    );
+    return true;
+  }
+
+  void _updatePositionList(_PositionItemType type, String displayValue) {
+    _positionItems.add(_PositionItem(type, displayValue));
+    setState(() {});
+  }
+
+  bool _isListening() => !(_positionStreamSubscription == null ||
+      _positionStreamSubscription!.isPaused);
+
+  Color _determineButtonColor() {
+    return _isListening() ? Colors.green : Colors.red;
+  }
+
+  void _toggleListening() {
+    if (_positionStreamSubscription == null) {
+      final positionStream = _geolocatorPlatform.getPositionStream();
+      _positionStreamSubscription = positionStream.handleError((error) {
+        _positionStreamSubscription?.cancel();
+        _positionStreamSubscription = null;
+      }).listen((position) => _updatePositionList(
+            _PositionItemType.position,
+            position.toString(),
+          ));
+      _positionStreamSubscription?.pause();
+    }
+
+    setState(() {
+      if (_positionStreamSubscription == null) {
+        return;
+      }
+
+      String statusDisplayValue;
+      if (_positionStreamSubscription!.isPaused) {
+        _positionStreamSubscription!.resume();
+        statusDisplayValue = 'resumed';
+      } else {
+        _positionStreamSubscription!.pause();
+        statusDisplayValue = 'paused';
+      }
+
+      _updatePositionList(
+        _PositionItemType.log,
+        'Listening for position updates $statusDisplayValue',
+      );
+    });
+  }
+
+  @override
+  void dispose() {
+    if (_positionStreamSubscription != null) {
+      _positionStreamSubscription!.cancel();
+      _positionStreamSubscription = null;
+    }
+
+    super.dispose();
+  }
+}
+
+enum _PositionItemType {
+  log,
+  position,
+}
+
+class _PositionItem {
+  _PositionItem(this.type, this.displayValue);
+
+  final _PositionItemType type;
+  final String displayValue;
+}